requirements.txt
# requirements.txt

absl-py==2.1.0
accelerate==1.1.1
ace_tools==0.0
ace_tools_open==0.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.11.4
aiosignal==1.3.1
altair==5.5.0
altgraph @ file:///AppleInternal/Library/BuildRoots/860631e9-c1c5-11ee-98ee-b6ef2fd8d87b/Library/Caches/com.apple.xbs/Sources/python3/altgraph-0.17.2-py2.py3-none-any.whl
annotated-types==0.7.0
anyio==4.4.0
appnope==0.1.4
arch==7.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astor==0.8.1
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==5.0.1
attrs==24.2.0
Babel==2.15.0
bayes==0.1.1
bayesian-optimization==2.0.0
bcrypt==4.2.1
beautifulsoup4==4.12.3
bert-score==0.3.13
bertopic==0.16.4
black==22.8.0
bleach==6.1.0
blinker==1.9.0
cachetools==5.5.0
captcha==0.6.0
certifi==2024.7.4
cffi==1.17.0
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
cmdstanpy==1.2.5
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.0
cryptography==44.0.0
cycler==0.12.1
datasets==3.1.0
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
distlib==0.3.9
distro==1.9.0
einops==0.8.0
entrypoints==0.4
et-xmlfile==1.1.0
evaluate==0.4.3
exceptiongroup==1.2.2
executing==2.0.1
extra-streamlit-components==0.1.71
Faker==33.1.0
fastjsonschema==2.20.0
favicon==0.7.0
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.54.1
fqdn==1.5.1
frozendict==2.4.6
frozenlist==1.5.0
fsspec==2024.9.0
future @ file:///AppleInternal/Library/BuildRoots/860631e9-c1c5-11ee-98ee-b6ef2fd8d87b/Library/Caches/com.apple.xbs/Sources/python3/future-0.18.2-py3-none-any.whl
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
grpcio==1.68.1
h11==0.14.0
h5py==3.12.1
hdbscan==0.8.39
holidays==0.62
htbuilder==0.7.0
html5lib==1.1
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.26.2
humanfriendly==10.0
identify==2.6.2
idna==3.7
importlib_metadata==8.2.0
importlib_resources==6.4.5
iniconfig==2.0.0
interchange==2021.0.4
ipykernel==6.29.5
ipython==8.18.1
ipywidgets==8.1.3
isoduration==20.11.0
isort==5.11.5
itables==2.2.3
jedi==0.19.1
Jinja2==3.1.4
jiter==0.7.1
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==7.4.9
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.11
keras==3.7.0
kiwisolver==1.4.7
libclang==18.1.1
lightgbm==4.5.0
llvmlite==0.43.0
lxml==5.3.0
macholib @ file:///AppleInternal/Library/BuildRoots/860631e9-c1c5-11ee-98ee-b6ef2fd8d87b/Library/Caches/com.apple.xbs/Sources/python3/macholib-1.15.2-py2.py3-none-any.whl
Markdown==3.7
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.4.1
monotonic==1.6
more-itertools==10.5.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
multitasking==0.0.11
mypy-extensions==1.0.0
namex==0.0.8
narwhals==1.14.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
neo4j==5.27.0
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.9.1
nodeenv==1.9.1
notebook==7.2.1
notebook_shim==0.2.4
NucleusApp==0.1.1b6
nucleusapp-telegram==0.2.2
NucleusUtils==0.1.11b1
numba==0.60.0
numpy==2.0.2
openai==1.54.4
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
opt_einsum==3.4.0
optimum==1.23.3
optree==0.13.1
overrides==7.7.0
packaging==24.1
paconvert @ file:///Users/ysomad/.vscode/extensions/baiducomate.comate-3.7.2/dist/comate-engine/plugins/paddle/assets/paconvert-2.5.0-py3-none-any.whl
pandas==2.2.3
pandas-datareader==0.10.0
pandocfilters==1.5.1
pansi==2024.11.0
parso==0.8.4
pathspec==0.12.1
patsy==1.0.1
pdfminer.six==20231228
pdfplumber==0.11.4
peewee==3.17.7
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.2.2
plotly==5.24.1
pluggy==1.5.0
portalocker==3.0.0
pre_commit==4.0.1
prettytable==3.12.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
propcache==0.2.0
prophet==1.1.6
protobuf==5.28.3
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py2neo==2021.2.4
pyaml==24.9.0
pyarrow==18.0.0
pybind11==2.13.6
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
pyecharts==2.0.7
Pygments==2.18.0
PyJWT==2.10.1
pylcs==0.1.1
pymdown-extensions==10.12
pynndescent==0.5.13
pyparsing==3.2.0
pypdfium2==4.30.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.15
PyQt5_sip==12.15.0
pyTelegramBotAPI==4.24.0
pytesseract==0.3.13
pytest==8.3.4
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytz==2024.2
PyYAML==6.0.2
pyzmq==24.0.1
qtconsole==5.5.2
QtPy==2.4.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rouge-score==0.1.2
rpds-py==0.20.0
sacrebleu==2.4.3
safetensors==0.4.5
scikit-learn==1.5.2
scikit-optimize==0.10.2
scipy==1.13.1
seaborn==0.13.2
Send2Trash==1.8.3
sentence-transformers==3.3.1
simplejson==3.19.3
six @ file:///AppleInternal/Library/BuildRoots/860631e9-c1c5-11ee-98ee-b6ef2fd8d87b/Library/Caches/com.apple.xbs/Sources/python3/six-1.15.0-py2.py3-none-any.whl
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.36
st-annotated-text==4.0.1
st-theme==1.2.3
stack-data==0.6.3
stanio==0.5.1
statsmodels==0.14.4
streamlit==1.40.1
streamlit-antd-components==0.3.2
streamlit-authenticator==0.4.1
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-echarts==0.4.0
streamlit-echarts5==1.0.0
streamlit-embedcode==0.1.2
streamlit-extras==0.5.0
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.9
streamlit-keyup==0.2.4
streamlit-on-Hover-tabs==1.0.1
streamlit-shadcn-ui==0.1.18
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
sympy==1.13.1
tabulate==0.9.0
TeleSocketClient==0.1.3b1
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.3.0
tokenizers==0.20.3
toml==0.10.2
tomli==2.0.1
torch==2.5.1
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
transformers==4.46.2
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.2
umap-learn==0.5.7
uri-template==1.3.0
urllib3==2.2.2
validators==0.34.0
vec2text==0.0.12
virtualenv==20.27.1
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
widgetsnbextension==4.0.11
wrapt==1.17.0
xgboost==2.1.2
xlrd==2.0.1
XlsxWriter==3.2.0
xxhash==3.5.0
yarl==1.17.2
yfinance==0.2.48
zipp==3.19.2
